Relocate billing hooks
