+++ src/components/admin/Billing/index.tsx
@@ -8,8 +8,8 @@ import AdminTabs from 'components/admin/Tabs';
 import MessageWithLink from 'components/content/MessageWithLink';
 import PageContainer from 'components/shared/PageContainer';
 import ProjectedCostsTable from 'components/tables/Billing';
-import useBillingCatalogStats from 'components/tables/Billing/useBillingCatalogStats';
 import { semiTransparentBackground } from 'context/Theme';
+import useBillingCatalogStats from 'hooks/billing/useBillingCatalogStats';
 import useBrowserTitle from 'hooks/useBrowserTitle';
 import useCombinedGrantsExt from 'hooks/useCombinedGrantsExt';
 import { useEffect } from 'react';
